Move stream options to dedicated header
